Revert "Fix endianness in header"
The previous behaviour was correct.  A width of 1900 pixels which is
0x00000640 in hex was stored on disk as:

00 00 06 40

That's BE order which is what we want.

Reading this uint32_t on an LE machine we get:

width = 0x40060000 and converting from BE to LE

we get the correct 0x640 value.

Reading on a BE machine we get the correct value
and the swap is a no-op.