Use stdin and stdout with 2ff
It always seemed odd to have a file-argument for 2ff, given all other
farbfeld conversion tools operate on stdin and stdout only.
Given I received numerous e-mails on this matter I've decided to change
the usage to lessen the confusion.

Maybe there's a way to do some magic with tee(1), but I didn't bother
just yet.
To ease the transition, the number of arguments are checked and a usage
printed if arguments are given.