Improve error-handling in the tools
We don't need the jumps, but rather pass a nice function pointer to
libpng. The jump in libjpg was also useless.
We check each fwrite-call so there's an early bailout in case the output
file is full.
Flushing at the end worked as well, but it took too long for complex
images. We don't want to "block" a pipe here and the approach in jpg2ff
was better.
The iHDR-read was useless. Rather use the get*-functions in libpng,
saves us 2 local variables as well.