Back to the roots.
Talking with even more people and weighing the benefits, I've made the
decision to remove color spaces from farbfeld again (as it currently
is handled in version 1) and recommend sRGB, not force it.
Thing is the following: We are not ready yet for this step. Neither the
people working with the images nor the hardware to display extended
colors. Using greater colorspaces also removes the "hackability" of
farbfeld.
As it was previously possible to easily create gradients in plain C,
you have to keep multiple things in mind with linear ProPhoto RGB.
The first thing is that not all colors are "real", and there are
imaginary colors. This doesn't happen with sRGB because all the colors
it describes are "real".
The second thing is the linear gamma curve, which makes the gradients
look perceptually inconsistent.

In the interest of creating a good and simple interchange format, we
can't only weigh in points of professionals. This little excursion has
shown that aiming for the 99% is still the way to go.
And as often as you could repeat that sRGB will meet its fate in the
future when screens become better, it is surprising how badly the
industry has caught up to it.
I also must honestly say that we can't solve the color space issue
using RGB and should look at other formats to explore (CIELUV, CIELAB),
which I will probably give a talk about at slcon3.

Before releasing version 2, my aim is to make the I/O a bit more
effective (as far as possible) so it's even easier to use.

I know people will be pissed off, but fuck it, I'm the maintainer!
Live with it.