Shellcheck 2ff fixes
I was inspired by the current discussion on dev@ to use shellcheck to
check my scripts and thought it might be a good choice to do this for
the 2ff script.
Not much had to be changed, because I was careful writing it, but still
it won't hurt to but $TMP in double quotes.