Refactor jpg-conversion-utilities
First of all, there was lots of libjpeg-specific cruft that just didn't
have any right to exist (METHODDEF(), strange typedefs, use of the
internal memory pool for no reason). This is gone now.

Additionally, we make use of the save and proven utility functions and
in general the code should be more well-separated now.

What is left to do is clear up the part where we mix the colors with the
mask.