Update TODO
The truncation issue is solved now by checking the return values. Maybe
for small images where the FILE-buffer is not flushed early enough we
can rethink that.
The utility functions were written to the extent it was desired.

I added a note not to forget to take a look at the part in the jpg-code
which does the color mixing.