Refactor invert.c in farbfeld.5
I noticed that it would be beneficial to release the invert.c code
listing under a very permissive license.
I like the style of the "Copy me if you can"-License, but thought
that 0BSD would make it even clearer that everyone can do whatever
he wants with this code.

The code itself was not bad beforehand, but lacked some elementary
features like checked flushing at the end and proper error messages.
I also reworked the data structures a bit to make it more appealing
and clearer where the "guts" of the code are (i.e. in invert()).