Bump version to 3
This is more or less a refactoring release, but with deep changes in the
tools that I was hoping to look into for a long time.
The codebase is in a very consistent state now, also thanks to the
introduction of a set of common utility-functions.

What really makes me think is the fact that it takes so many iterations
and a high level of detail to get the library handling and I/O right.
It just makes you wonder how much software is out there that is full
of little subtle bugs that might blow up in your face some day.

Thanks for all the feedback!