Makefile-workaround for OpenBSD
Thanks Hiltjo for reporting this!

For some reason, OpenBSD make does not add $(BIN).o to the dependency list
when it sees the $(BIN) target rule. It does so however when it does the
conversion from .c to .o.
This behaviour is unique for OpenBSD make as far as I've seen, and for
the time being, this workaround will manually add the object to the
dependency list.
This is not POSIX compliant and in GNU make it just evaluates to an
empty string, but it works until this has been investigated further.