Combine all license matters in LICENSE
We use some OpenBSD code in util.c (namely by Ted Unangst, Todd C.
Miller and Otto Moerbeek), which is licensed under the ISC license.

To make it clearer and make it easier to see what farbfeld really is licensed
under, we include these authors in the LICENSE file and remove the
explicit headers from util.c. While at it, we also remove superfluous
includes and shove them to the top.