arg.h: Do not modify the array elements of argv[]
The C99 standard explicitly allows to modify argc, argv, but leaves it
open what happens if you modify the content of argv. Under OpenBSD, this
actually has an effect on how the program is listed (e.g. in ps). To
prevent this, we just add a counter variable and use that for iteration.

While at it, this commit also includes a few style changes.

Thanks Hiltjo for reporting this!