Add efread() and efwrite()
Given we have a util.c anyway it does not make any sense to duplicate
code here. This way, error handling and output is in one place and one
does not have to change things in multiple different files. This also
reduces code duplication by a lot.

It also solves an older bug where the error reporting was not on point:

   $ echo "farb" | ff2png
   ff2png: fread: Success

(It also lacked a newline)

Now it properly reports

   $ echo "farb" | ff2png
   ff2png: fread: Unexpected end of file

I also fixed some other minor details, for instance that all error
messages should begin with a capital letter.