Bump version to 4
Foremost this release finally includes a fix for the little mistake I made in
the Makefile regarding the order of arguments passed to cc which package
managers had to fix with a small patch. It now also compiles on OpenBSD
without problems.

Apart from that, I made error reporting much more consistent by putting
everything into util.c, which makes the code much more readable, and
cleared up all LICENSE matters and minor details in the build system.

At this point, farbfeld is very stable and unless another tool is added
the codebase in general is very low on the Kelvin scale right now.

Thanks for all the feedback!